+++ src/platform_impl/ios/monitor.rs
@@ -2,11 +2,11 @@
 
 use std::{
     collections::{BTreeSet, VecDeque},
-    fmt,
-    ops::{Deref, DerefMut},
+    fmt, hash, ptr,
 };
 
-use icrate::Foundation::{MainThreadMarker, NSInteger};
+use icrate::Foundation::{MainThreadBound, MainThreadMarker, NSInteger};
+use objc2::mutability::IsRetainable;
 use objc2::rc::Id;
 
 use super::uikit::{UIScreen, UIScreenMode};
@@ -16,32 +16,59 @@ use crate::{
     platform_impl::platform::app_state,
 };
 
-// TODO(madsmtm): Remove or refactor this
-#[derive(Debug, PartialEq, Eq, Hash, Clone)]
-pub(crate) struct ScreenModeSendSync(pub(crate) Id<UIScreenMode>);
+// Workaround for `MainThreadBound` implementing almost no traits
+#[derive(Debug)]
+struct MainThreadBoundDelegateImpls<T>(MainThreadBound<Id<T>>);
 
-unsafe impl Send for ScreenModeSendSync {}
-unsafe impl Sync for ScreenModeSendSync {}
+impl<T: IsRetainable> Clone for MainThreadBoundDelegateImpls<T> {
+    fn clone(&self) -> Self {
+        Self(
+            self.0
+                .get_on_main(|inner, mtm| MainThreadBound::new(Id::clone(inner), mtm)),
+        )
+    }
+}
+
+impl<T: IsRetainable> hash::Hash for MainThreadBoundDelegateImpls<T> {
+    fn hash<H: hash::Hasher>(&self, state: &mut H) {
+        // SAFETY: Marker only used to get the pointer
+        let mtm = unsafe { MainThreadMarker::new_unchecked() };
+        Id::as_ptr(self.0.get(mtm)).hash(state);
+    }
+}
+
+impl<T: IsRetainable> PartialEq for MainThreadBoundDelegateImpls<T> {
+    fn eq(&self, other: &Self) -> bool {
+        // SAFETY: Marker only used to get the pointer
+        let mtm = unsafe { MainThreadMarker::new_unchecked() };
+        Id::as_ptr(self.0.get(mtm)) == Id::as_ptr(other.0.get(mtm))
+    }
+}
+
+impl<T: IsRetainable> Eq for MainThreadBoundDelegateImpls<T> {}
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
 pub struct VideoMode {
     pub(crate) size: (u32, u32),
     pub(crate) bit_depth: u16,
     pub(crate) refresh_rate_millihertz: u32,
-    pub(crate) screen_mode: ScreenModeSendSync,
+    screen_mode: MainThreadBoundDelegateImpls<UIScreenMode>,
     pub(crate) monitor: MonitorHandle,
 }
 
 impl VideoMode {
-    fn new(uiscreen: Id<UIScreen>, screen_mode: Id<UIScreenMode>) -> VideoMode {
-        assert_main_thread!("`VideoMode` can only be created on the main thread on iOS");
+    fn new(
+        uiscreen: Id<UIScreen>,
+        screen_mode: Id<UIScreenMode>,
+        mtm: MainThreadMarker,
+    ) -> VideoMode {
         let refresh_rate_millihertz = refresh_rate_millihertz(&uiscreen);
         let size = screen_mode.size();
         VideoMode {
             size: (size.width as u32, size.height as u32),
             bit_depth: 32,
             refresh_rate_millihertz,
-            screen_mode: ScreenModeSendSync(screen_mode),
+            screen_mode: MainThreadBoundDelegateImpls(MainThreadBound::new(screen_mode, mtm)),
             monitor: MonitorHandle::new(uiscreen),
         }
     }
@@ -61,17 +88,39 @@ impl VideoMode {
     pub fn monitor(&self) -> MonitorHandle {
         self.monitor.clone()
     }
+
+    pub(super) fn screen_mode(&self, mtm: MainThreadMarker) -> &Id<UIScreenMode> {
+        self.screen_mode.0.get(mtm)
+    }
 }
 
-#[derive(Clone, PartialEq, Eq, Hash)]
-pub struct Inner {
-    uiscreen: Id<UIScreen>,
-}
-
-#[derive(Clone, PartialEq, Eq, Hash)]
 pub struct MonitorHandle {
-    inner: Inner,
+    ui_screen: MainThreadBound<Id<UIScreen>>,
 }
+
+impl Clone for MonitorHandle {
+    fn clone(&self) -> Self {
+        Self {
+            ui_screen: self
+                .ui_screen
+                .get_on_main(|inner, mtm| MainThreadBound::new(inner.clone(), mtm)),
+        }
+    }
+}
+
+impl hash::Hash for MonitorHandle {
+    fn hash<H: hash::Hasher>(&self, state: &mut H) {
+        (self as *const Self).hash(state);
+    }
+}
+
+impl PartialEq for MonitorHandle {
+    fn eq(&self, other: &Self) -> bool {
+        ptr::eq(self, other)
+    }
+}
+
+impl Eq for MonitorHandle {}
 
 impl PartialOrd for MonitorHandle {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
@@ -83,31 +132,6 @@ impl Ord for MonitorHandle {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
         // TODO: Make a better ordering
         (self as *const Self).cmp(&(other as *const Self))
-    }
-}
-
-impl Deref for MonitorHandle {
-    type Target = Inner;
-
-    fn deref(&self) -> &Inner {
-        assert_main_thread!("`MonitorHandle` methods can only be run on the main thread on iOS");
-        &self.inner
-    }
-}
-
-impl DerefMut for MonitorHandle {
-    fn deref_mut(&mut self) -> &mut Inner {
-        assert_main_thread!("`MonitorHandle` methods can only be run on the main thread on iOS");
-        &mut self.inner
-    }
-}
-
-unsafe impl Send for MonitorHandle {}
-unsafe impl Sync for MonitorHandle {}
-
-impl Drop for MonitorHandle {
-    fn drop(&mut self) {
-        assert_main_thread!("`MonitorHandle` can only be dropped on the main thread on iOS");
     }
 }
 
@@ -135,59 +159,80 @@ impl fmt::Debug for MonitorHandle {
 }
 
 impl MonitorHandle {
-    pub(crate) fn new(uiscreen: Id<UIScreen>) -> Self {
-        assert_main_thread!("`MonitorHandle` can only be created on the main thread on iOS");
+    pub(crate) fn new(ui_screen: Id<UIScreen>) -> Self {
+        // Holding `Id<UIScreen>` implies we're on the main thread.
+        let mtm = MainThreadMarker::new().unwrap();
         Self {
-            inner: Inner { uiscreen },
+            ui_screen: MainThreadBound::new(ui_screen, mtm),
         }
     }
-}
 
-impl Inner {
     pub fn name(&self) -> Option<String> {
-        let main = UIScreen::main(MainThreadMarker::new().unwrap());
-        if self.uiscreen == main {
-            Some("Primary".to_string())
-        } else if self.uiscreen == main.mirroredScreen() {
-            Some("Mirrored".to_string())
-        } else {
-            UIScreen::screens(MainThreadMarker::new().unwrap())
-                .iter()
-                .position(|rhs| rhs == &*self.uiscreen)
-                .map(|idx| idx.to_string())
-        }
+        self.ui_screen.get_on_main(|ui_screen, mtm| {
+            let main = UIScreen::main(mtm);
+            if *ui_screen == main {
+                Some("Primary".to_string())
+            } else if *ui_screen == main.mirroredScreen() {
+                Some("Mirrored".to_string())
+            } else {
+                UIScreen::screens(mtm)
+                    .iter()
+                    .position(|rhs| rhs == &**ui_screen)
+                    .map(|idx| idx.to_string())
+            }
+        })
     }
 
     pub fn size(&self) -> PhysicalSize<u32> {
-        let bounds = self.uiscreen.nativeBounds();
+        let bounds = self
+            .ui_screen
+            .get_on_main(|ui_screen, _| ui_screen.nativeBounds());
         PhysicalSize::new(bounds.size.width as u32, bounds.size.height as u32)
     }
 
     pub fn position(&self) -> PhysicalPosition<i32> {
-        let bounds = self.uiscreen.nativeBounds();
+        let bounds = self
+            .ui_screen
+            .get_on_main(|ui_screen, _| ui_screen.nativeBounds());
         (bounds.origin.x as f64, bounds.origin.y as f64).into()
     }
 
     pub fn scale_factor(&self) -> f64 {
-        self.uiscreen.nativeScale() as f64
+        self.ui_screen
+            .get_on_main(|ui_screen, _| ui_screen.nativeScale()) as f64
     }
 
     pub fn refresh_rate_millihertz(&self) -> Option<u32> {
-        Some(refresh_rate_millihertz(&self.uiscreen))
+        Some(
+            self.ui_screen
+                .get_on_main(|ui_screen, _| refresh_rate_millihertz(ui_screen)),
+        )
     }
 
     pub fn video_modes(&self) -> impl Iterator<Item = VideoMode> {
-        // Use Ord impl of RootVideoMode
-        let modes: BTreeSet<_> = self
-            .uiscreen
-            .availableModes()
-            .into_iter()
-            .map(|mode| RootVideoMode {
-                video_mode: VideoMode::new(self.uiscreen.clone(), mode),
-            })
-            .collect();
+        self.ui_screen.get_on_main(|ui_screen, mtm| {
+            // Use Ord impl of RootVideoMode
 
-        modes.into_iter().map(|mode| mode.video_mode)
+            let modes: BTreeSet<_> = ui_screen
+                .availableModes()
+                .into_iter()
+                .map(|mode| RootVideoMode {
+                    video_mode: VideoMode::new(ui_screen.clone(), mode, mtm),
+                })
+                .collect();
+
+            modes.into_iter().map(|mode| mode.video_mode)
+        })
+    }
+
+    pub(crate) fn ui_screen(&self, mtm: MainThreadMarker) -> &Id<UIScreen> {
+        self.ui_screen.get(mtm)
+    }
+
+    pub fn preferred_video_mode(&self) -> VideoMode {
+        self.ui_screen.get_on_main(|ui_screen, mtm| {
+            VideoMode::new(ui_screen.clone(), ui_screen.preferredMode().unwrap(), mtm)
+        })
     }
 }
 
@@ -213,20 +258,6 @@ fn refresh_rate_millihertz(uiscreen: &UIScreen) -> u32 {
     };
 
     refresh_rate_millihertz as u32 * 1000
-}
-
-// MonitorHandleExtIOS
-impl Inner {
-    pub(crate) fn ui_screen(&self) -> &Id<UIScreen> {
-        &self.uiscreen
-    }
-
-    pub fn preferred_video_mode(&self) -> VideoMode {
-        VideoMode::new(
-            self.uiscreen.clone(),
-            self.uiscreen.preferredMode().unwrap(),
-        )
-    }
 }
 
 pub fn uiscreens(mtm: MainThreadMarker) -> VecDeque<MonitorHandle> {